Convert Google Docs clipboard images to JPEG
